+++ docs/CHANGELOG.md
@@ -6,6 +6,20 @@ This is the Envlang alpha development version. Expect large API changes and sudd
 
 This version adds a bunch of operators as well as implements operator precedence.
 
+### Version 0.6.1
+
+#### Major changes
+
+- Added the symbol enum `LogicalOperators`. It enumerates three Boolean-logical operators: AND, OR, and NOT.
+- Added the num variant `Operators::Logical(LogicalOperators)`.
+- The lexer now recognises and lexes the unary logical operator NOT (`!`) and the binary logical operators AND (`&`) and OR (`|`).
+- The parser now recognises and parses binary logical operations (AND and NOT).
+
+#### Minor changes
+
+- Added tests for unary and binary logical operator lexing.
+- Added tests for binary logical operator parsing.
+
 ### Version 0.6.0
 
 #### Major changes
@@ -17,6 +31,7 @@ This version adds a bunch of operators as well as implements operator precedence
 #### Minor changes
 
 - Added tests for comparison operator lexing.
+- Added tests for comparison operator parsing.
 
 ## Version 0.5